Moved the unit testing work under the unit subdirectory to tidy some things up.
