+++ tests/unit/config/phpunit.xml.dist
@@ -11,7 +11,7 @@
         >
     <testsuites>
         <testsuite name="Blocklyduino Test Suite">
-            <directory>./tests/</directory>
+            <directory>../</directory>
         </testsuite>
     </testsuites>
 </phpunit>
